Change InputManager to be a ScriptableObject
Change the InputManager script to be a ScriptableObject instead of a
MonoBehaviour. Since there should only ever be one instance of InputManager, it
is a singleton and therefore makes sense to be a ScriptableObject. Also, this
will allow it to be referenced in the Unity Editor. Any Component that needs
to access the InputManager singleton just has to have a reference to the
PlayerInputManager ScriptableObject asset dragged to one of its fields and
Unity will take care of connecting them. This removes the need to do this in
code and provides visual debugging.