Implement a Finite State Machine for the Game
Implement a finite state machine for the game as a whole. This state machine
will keep track of what scenes are currently being displayed such as the main
menu or the game. All game states will be created as GameState
ScriptableObjects which can be easily dragged around in the editor in order to
set the States for various scripts.