Add the Detection of Mouse Input and Add a Script to Respond to it
Add the detection of mouse input that gets the position of the mouse in world
space. Also create a script that can be attached to a GameObject that will
allow it to respond to the location of the mouse and move to it. This will
allow GameObjects to move towards the location of the mouse which is what will
be the main method of movement in the final game.